feat(p2p): add comments for examples
